requirements.txt
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Faker==8.1.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.4.11
text-unidecode==1.3
Werkzeug==1.0.1
WTForms==2.3.3
pymysql
cryptography
boto3
cx-Oracle==8.1.0
azure-storage-blob